adfs

+++ Monopoly/src/actions/GameAction.java
@@ -43,7 +43,7 @@ public class GameAction extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
 		//Session Daten abfragen
-		int gameId = (int) request.getSession().getAttribute(TextKeys.userGameId);
+		long gameId = (long) request.getSession().getAttribute(TextKeys.userGameId);
 		int playerId = (int) request.getSession().getAttribute(TextKeys.playerId);
 		
 		//Spiel in HashMap finden
